Exercise5
I defined a dictionary where each month is represented as a number and its corresponding number of days it has. Variable of days in February in leap year is made for later use. I then ask user for a month in the range of 1-12, if else statement used as well as nested if else statement in order to validate user's input and whether it is leap year or not in order to give out the correct number of days in February as well as other months.