Exercise7
I created a loop using for i in range to count up from 0 to 50 in increments of 1; another loop that counts down from 50 to 0 in decrements of 1; a loop that counts up from 30 to 50 in increments of 1; a loop that counts down from 50 to 10 in decrements of 2, and finally a loop that counts up from 100 to 200 in increments of 5. each number in the  loop is separated with "." and each loop is separated by two empty lines.